pyproject.toml
[project]
name = "protobuf-grpc-distutils-example"
dynamic = ["version"]
dependencies = [
    "grpcio",
    "protobuf",
    
]

[build-system]
requires = [
    "setuptools>=61.0", 
    "wheel",
    "setuptools_scm[toml]>=6.0",
    "protobuf_grpc_distutils @ git+https://github.com/idex-biometrics/protobuf-grpc-distutils@main#egg=protobuf_grpc_distutils"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "helloworld.__version__"}

[tool.setuptools_scm]
write_to = "helloworld/_version.py"

[project.scripts]
hello_client = "helloworld.client:run"
hello_server = "helloworld.server:serve"

[tool.setuptools.packages]
find = {}
